Add example of seas.virginia.edu Beamer theme
This changeset adds two images of a presentation styled with the
seas.virginia.edu theme to the theme's documentation. It makes is
easier to see the immediate effect of changes to the theme.

The example presentation reuses the same example presentation as the
themes shown in Beamer's documentation. beamerthemeexamplebase.tex is
copied (unmodified) from Beamer's official repository
(https://goo.gl/Psoxm8), changeset 3dd33d7.