Add examples of Beamer themes
This changeset adds two images of a presentation styled with the theme
to the theme's documentation. It makes it easier to see the theme and
the immediate effect of any changes to the theme.

The testcolor macro is added to the theme-doc package to replace
duplicate definitions in the bluejay and wahoo color themes.