Use logo key-value options in derived themes
This changeset uses the key-value options defined in the logo outer
theme to control the formatting of derived themes. The jhu.edu theme
serves as an example of the changes when exposing these options to
users.

This changeset also bundles a variety of additional modifications to
standardize color palettes, etc. Unraveling these modifications from
those pertaining to a consistent interpretation of options in the logo
outer theme was deemed too difficult to break into separate
changesets.

Using pgfkeys to handle package options has several drawbacks that
were exposed when working on this changeset. In particular, the
default options in some cases affect global state. A good example is
loading the `base' theme: ideally the base theme is loaded once, but
executing the default options causes the theme to be loaded
immediately rather than deferring until specified by the user (and
using the defaults as a fallback). The workaround for this behavior
has its own side effects, and it may be best to abandon this line of
development in favor of a different implementation.