Create CTAN distribution archives for packages
This change creates a distribution archive for each package when
`make` is executed in the root directory of the repository. The
distribution archive for each packages comprises its documented LaTeX
(.dtx), installer (.ins), and documentation.

For distribution, CTAN prescribes the structure of packages. In
particular, all files should be inside a directory that is inside a
ZIP archive. That is, the contents of the distribution archive should
be as follows:

  my-package.zip:
    my-package/
      README
      my-package.dtx
      my-package.ins
      my-package.pdf

This change follows that recommendation.

CTAN instructions: https://ctan.org/help/upload-pkg