Matrix build workflow for TeX Live versions
This change defines multiple operating system versions (Debian 10 and
11) to identify issues stemming from different TeX Live versions (2018
and 2020 respectively), including LaTeX kernel changes.

The dependencies installed for each Debian release are customized to
handle transitional packages and packages that are only required for a
particular release, such as python-is-python3 to ensure that Debian 11
("bullseye") has a `python` executable (instead of only a `python3`
executable).